app/requirements/base.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
sqlparse==0.4.4
psycopg2
django-allauth==0.61.1
dj-rest-auth[with_social]==6.0.0
drf-yasg